KernelInfo::getMatchedKernelRequirements takes kernel FCM version
Add kernel FCM version as an arg to getMatchedKernelRequirements.
The legacy behavior (devices with kernel FCM version UNSPECIFIED) is
preserved.

New behavior (devices with kernel FCM version set) is
implemented, but matrices don't have the correct <kernel> tags for
these devices yet. This is fixed and tested in the next CL.

Test: libvintf_test
Test: vintf_object_test

Bug: 139309488

Change-Id: I6f0d8cbb1424116d6d07fafa6573b29d6fb720fe