assemble_vintf: accept kernel config name != android-base.config
Accept additional file names other than android-base.config.
As long as the name does not start with 'android-base-', put it in
common configs.

This allows additional "extra" common configs as a supplement to
android-base.config. Specifically, for non-debuggable builds,
CONFIG_DEBUG_FS must not be set. An additional non-debuggable.config is
added on non-debuggable builds.

Test: build user and userdebug
Bug: 152633647
Change-Id: Idd9de8c7638f27a741e0925051b5e8128a01497c