Inject empty <kernel> tag if missing.
assemble_vintf should infer kernel FCM version when it is
missing (for devices with target FCM version >= R), but it
only did it when an empty <kernel> tag is present. Fix by
injecting the empty kernel tag if it is not there.

Test: delete <kernel> tag from CF and m check-vintf-all
Bug: 155604363
Change-Id: I216a5de6a8613c3ca270365b9ab536f6862fc8ad
Merged-In: I216a5de6a8613c3ca270365b9ab536f6862fc8ad