Remove skipCache for RuntimeInfo
The argument was created for vintf_object_tests before VintfObject is
instantiated as an object. Now that an empty VintfObject is created
before each test, this argument can be safely dropped.

Also, fix VintfObjectRuntimeInfoTest.GetRuntimeInfo because that test
tests the behavior of skipCache. Now that it is removed, just test that
the information is fetched and subsequent calls won't fetch the
information again.

Test: unit tests
Bug: 110855270
Change-Id: Ib56d413e0cf188225595ae5f4da1aedcff1a14ce