Make checkvintf warn on missing HAL entries in matrices.
When building OTA packages, we may be using hidl/aidl_metadata and
exempt list that is too new for the target files package we are dealing
with. Hence, make this a warning until the problem is resolved.

When building, even though check_vintf throws a warning,
check_vintf --check-one still throws an error. So at build time, the
missing HALs will still cause a build failure.

Bug: 168453016
Bug: 110261831

Test: manual on an incompatible package

Change-Id: I31d23debe62ac62f4a3e4d5846942bde04cf7d01