Change fake/default AIDL version to (size_max, 1).
Previously it is (size_max, size_max), and it is okay because the
version number is not used anywhere, but only as a trick to get
forEachInstance working.

Now that we are adding versioning to AIDL, change the minor version to 0
instead. The minor version in VINTF is equivalent to the AIDL version
because all AIDL interfaces are backwards compatible, and
non-backwards-compatible changes are reflected in the change in the
package name.

Test: libvintf_test
Test: vintf_object_test
Bug: 150034914

Change-Id: I3356ed5ce818ecd4330628e8930aa406c2cca2da