Add <version> to AIDL HALs in manifest.
Unlike HIDL HALs, the <version> tag for AIDL HALs is a single
integer:

    <version>0</version>

The version of an AIDL HAL is similar to the minor version of
a HIDL HAL, because all AIDL HALs are backwards compatible.

Note that in libvintf@3.0 and before, <version> tags in AIDL HALs
yields undefined behavior. Previously, if they exist:
- If it has the format x.y, libvintf will emit a warning and ignore it
- If it has any other formats, libvintf will refuse to parse the whole
  file.

Since the format of <version> tag of AIDL HALs effectively changes, the
metaversion of libvintf has been bumped to 4.0.

Test: libvintf_test
Test: vintf_object_test
Bug: 150034914

Change-Id: I47b237df6bcdad3610ff4b731d4543b048efc566