Add <version> to AIDL HALs in matrix.
Unlike HIDL HALs, the <version> tag for AIDL HALs does not contain
the major version:

    <version>1</version>

OR

    <version>1-3</version>

The version of an AIDL HAL is similar to the minor version of a HIDL
HAL, because all AIDL HALs are backwards compatible.

Note that in libvintf@3.0 and before, <version> tags in AIDL HALs
yields undefined behavior. Previously, if they exist:
- If it has the format x.y, libvintf will emit a warning and ignore it
- If it has any other formats, libvintf will refuse to parse the whole
  file.

Since the format of <version> tag of AIDL HALs effectively changes, the
metaversion of libvintf has been bumped to 4.0.

Test: libvintf_test
Bug: 175821740

Change-Id: If7d300cdc193dcfaa7e95429fc7fbd593945633f