Disallow specifying version in <fqname> for AIDL HAL.
In libvintf@3.0 and before, <fqname> tags in AIDL HALs
with versions yields undefined behavior. To be more specific, the
old behavior is that if an AIDL HAL specifies:
    <fqname>@100.100::IFoo/default</fqname>
... then libvintf silently replaces the 100.100 version with
kFakeAidlVersion.

This change explicitly disallows specifiying version
in the <fqname> of an AIDL HAL.

Also update references to kFakeAidlVersion to kDefaultAidlVersion
in code nearby, even though kDefaultAidlVersion is used as a fake
version in this case.

Test: libvintf_test
Test: vintf_object_test
Bug: 150034914

Change-Id: I5ee2ac305fdfa36dce22e3c416e5bd7a13bebe19