ManifestHal no longer stores <interface>
All <interface>x<instance> tuples are now transformed
into <fqname> before being stored in memory.

With this change, <version>x<interface>x<instance> are
converted into <fqname> when the XML is parsed, not when
the XML is generated, because <interface> is no longer stored.
This means the logic to parse <hal> for a manifest is
reworked, and checks are performed when the XML is parsed.

Tests are added to ensure the change is backwards compatible;
that is, existing XML files with <version> x <interface> x <instance>
are parsed correctly. Only the final output is affected
that <version> x <interface> x <instance> is no longer produced.
Specifically:

- For AIDL HALs, <interface> x <instance> is converted to <fqname>;
  <version> is kept because <fqname> does not contain a version.
- For HIDL HALs and native HALs with <version> x <interface> x <instance>,
  the cross product list is converted to <fqname>.
- For HIDL HALs and native HALs with <version> but no <interface>,
  <version> is kept to represent the legacy behavior of these <hal>s.

With this change, the ManifestHal.interfaces field
is no longer used. Delete code associated with it.

- Also delete the constructor for ManifestHal that
  was only used in tests. In tests, assign
  fields to a default-constructed ManifestHal directly.

Test: libvintf_test
Test: vintf_object_test
Bug: 148808037
Change-Id: I1f03bf759151c8435bf88e0b08b1b0432ab76229