Do not allow <hal> without any <fqname> or <instance>
For manifests >= libvintf 6.0, require HIDL and AIDL HALs
to always have at least one <fqname> or <instance>.

Though it is technically okay to not have any instances,
this usually indicates that the author has forgotten
to put in any instances. If this is intentional, the author
should delete the <hal> tag completely.

Test: libvintf_test
Bug: 148808037
Change-Id: I3606830d74ff81b27f4c65670b65dfa3247665b6